requirements.txt
deepgram-sdk
websockets